Arreglos relacionados con la seguridad

+++ Practica3/pagina/administracion/pelicula/borrarPelicula.php
@@ -2,15 +2,17 @@
 
     require_once('../../../includes/config.php');
     
-    if (isset($_POST['id'])) {
-        $pelicula = es\ucm\fdi\aw\pelicula::buscar($_POST['id']);
-        if ($pelicula && es\ucm\fdi\aw\pelicula::borrar($_POST['id'])) {
-            unlink(RUTA_RAIZ . RUTA_PSTR . "/" . $pelicula->getRutaPoster()); 
-            unlink(RUTA_RAIZ . RUTA_TRL . "/" . $pelicula->getRutaTrailer());
-            header('location: ' . RUTA_APP . RUTA_ADMN);
+    if (comprobarPermisos($_SESSION["esAdmin"])) {
+        if (isset($_POST['id'])) {
+            $pelicula = es\ucm\fdi\aw\pelicula::buscar($_POST['id']);
+            if ($pelicula && es\ucm\fdi\aw\pelicula::borrar($_POST['id'])) {
+                unlink(RUTA_RAIZ . RUTA_PSTR . "/" . $pelicula->getRutaPoster()); 
+                unlink(RUTA_RAIZ . RUTA_TRL . "/" . $pelicula->getRutaTrailer());
+                header('location: ' . RUTA_APP . RUTA_ADMN);
+            }
         }
-    }
-    else {
-        echo 'Error al borrar la película';
-        exit();
+        else {
+            echo 'Error al borrar la película';
+            exit();
+        }
     }
